Adding "unsafe" option for reading dted
Many dted0 files I've come across have bad or empty fields in various
headers (especially edition and dates) and will not be able to be read
with the current implementation. Also, many have void data, which throws
a warning, but if you know the file and don't want to see the warning,
this adds a way to handle that as well.